hdmirx: update vrrmax range [1/1]
PD#SWPL-235689

Problem:
vrrmax only has 8 bits and cannot declare 360Hz

Solution:
vrrmax changed to u32

Verify:
t6x

+++ drivers/media/vin/tvin/hdmirx/hdmi_rx_drv.h
@@ -993,7 +993,7 @@ struct rx_info_s {
 	bool pip_on;
 	bool hdr10p_en;
 	u8 vrr_min;
-	u8 vrr_max;
+	u32 vrr_max;
 	u32 arc_port;
 	bool arc_5vsts;
 	unsigned long timestamp;
